requirements.txt
django==5.1.3
psycopg2-binary
django-environ
djangorestframework
django-filter